requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20230430
Django==5.0
django-admin-interface==0.29.4
django-allauth==65.0.2
django-celery-beat==2.7.0
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-import-export==4.1.1
django-jazzmin==3.0.0
django-simple-history==3.7.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
kombu==5.4.2
numpy==2.2.2
opencv-contrib-python==4.11.0.86
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
text-unidecode==1.3
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
